The semantics now use namespaces as unique symbol identifiers.
This harmonises a lot.
A simplification of the semantic symbols classes is included in this commit. While the simplification is a good thing, it is unfortunately included in this commit as seperating both changes seemed impossible.
To not further complicate the connector contexts, a simplification and standardisation of the connector context was done.
There is a sideeffect of this change that type inference on module variables and fields does not work anymore. This is the case because it was necessary to prescan the variable and field declarations (like with functions), so it was possible to connect the module before connecting the variable/field initialisers (which include expressions which need the module context).